Update dictionaryproj.py
So I did a lottt. My main focus was trying to make the schedule correctly, and I did that and added comments explaining the processes. Now I need to focus on adding, changing, and removeing the classes, and I got stuck now. So I am going to bed/ (I realized the method I used to make the schedules may not work for adding in the classes... oh no)