Update dictionaryproj.py
Today I worked on fixing the add_class function, and it worked! so I am happy. I realized that I have to make it so that is only possible if there is a free block, so ill do that later, because it works more like changing a class now. I started the change_class function and need to tweak it, and i still need to do the drop class function. I want to make the program restartable with saving the progress if the user wants to change more things in their schedule. I need to do that all tonight