Update dictionaryproj.py
I still need to fix the add_class thing a bit because of the part where it's like "You already have a class in here" (it ends the add_class func but I want it to go back. I could also just leave it, but it would be nice if I could go back to asking the question.

 I did the change_class and drop_class function. I made a restart function and made it able to save the past schedule until the user trashes the terminal. yay